+++ src/ext_widgets/scroll_view/scroll_view.h
@@ -186,6 +186,11 @@ typedef struct _scroll_view_t {
  */
 
 /**
+ * @event {event_t} EVT_PAGE_CHANGING
+ * 页面正在改变。
+ */
+
+/**
  * @method scroll_view_create
  * 创建scroll_view对象
  * @annotation ["constructor", "scriptable"]